refactor: api controller와 일반 controller 분리
